backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
PyMySQL>=1.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
